+++ app/src/main/java/com/orange/ods/app/ui/components/tabs/ComponentTabs.kt
@@ -18,6 +18,8 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.pager.HorizontalPager
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableIntStateOf
@@ -32,7 +34,10 @@ import com.orange.ods.app.ui.TabsConfiguration
 import com.orange.ods.app.ui.components.Variant
 import com.orange.ods.app.ui.components.utilities.ComponentCountRow
 import com.orange.ods.app.ui.components.utilities.ComponentCustomizationBottomSheetScaffold
+import com.orange.ods.app.ui.utilities.code.CodeImplementationColumn
+import com.orange.ods.app.ui.utilities.code.FunctionCallCode
 import com.orange.ods.app.ui.utilities.composable.Subtitle
+import com.orange.ods.compose.OdsComposable
 import com.orange.ods.compose.component.chip.OdsChoiceChipsFlowRow
 import com.orange.ods.compose.component.listitem.OdsListItem
 import com.orange.ods.compose.component.tab.OdsTabRow
@@ -65,7 +70,7 @@ fun ComponentTabs(variant: Variant) {
     with(tabsCustomizationState) {
         val appBarManager = LocalAppBarManager.current
         appBarManager.updateAppBarTabs(
-            TabsConfiguration(scrollableTabs, tabs, pagerState, tabsIconPosition.value, tabIconEnabled.value, tabTextEnabled.value)
+            TabsConfiguration(scrollableTabs, tabs, pagerState, tabIconPosition.value, tabIconEnabled.value, tabTextEnabled.value)
         )
 
         ComponentCustomizationBottomSheetScaffold(
@@ -81,14 +86,18 @@ fun ComponentTabs(variant: Variant) {
                 )
                 Subtitle(textRes = R.string.component_tabs_icon_position, horizontalPadding = true)
                 OdsChoiceChipsFlowRow(
-                    selectedChoiceChipIndex = OdsTabRow.Tab.Icon.Position.entries.indexOf(tabsIconPosition.value),
+                    selectedChoiceChipIndex = OdsTabRow.Tab.Icon.Position.entries.indexOf(tabIconPosition.value),
                     modifier = Modifier.padding(horizontal = dimensionResource(id = com.orange.ods.R.dimen.spacing_m)),
-                    choiceChips = OdsTabRow.Tab.Icon.Position.entries.map { tabsIconPosition ->
-                        val textResId = when (tabsIconPosition) {
+                    choiceChips = OdsTabRow.Tab.Icon.Position.entries.map { tabIconPosition ->
+                        val textResId = when (tabIconPosition) {
                             OdsTabRow.Tab.Icon.Position.Top -> R.string.component_tabs_icon_position_top
                             OdsTabRow.Tab.Icon.Position.Leading -> R.string.component_tabs_icon_position_leading
                         }
-                        OdsChoiceChipsFlowRow.ChoiceChip(stringResource(id = textResId), { this.tabsIconPosition.value = tabsIconPosition }, isTabsIconPositionEnabled)
+                        OdsChoiceChipsFlowRow.ChoiceChip(
+                            stringResource(id = textResId),
+                            { this.tabIconPosition.value = tabIconPosition },
+                            isTabIconPositionEnabled
+                        )
                     }
                 )
 
@@ -101,23 +110,52 @@ fun ComponentTabs(variant: Variant) {
                     minCount = tabCountMin,
                     maxCount = tabCountMax
                 )
-            }) {
+            }
+        ) {
+            HorizontalPager(state = pagerState) { pageIndex ->
+                val modifier = if (pageIndex == 0) {
+                    Modifier
+                        .fillMaxSize()
+                        .verticalScroll(rememberScrollState())
+                        .padding(top = dimensionResource(id = com.orange.ods.R.dimen.screen_vertical_margin))
+                } else {
+                    Modifier
+                        .fillMaxSize()
+                        .wrapContentSize(Alignment.Center)
+                }
 
-            HorizontalPager(state = pagerState) { page ->
-                val textResId = tabs[page].textResId
-                TabsPagerContentScreen(stringResource(id = textResId))
+                Column(modifier = modifier) {
+                    if (pageIndex == 0) {
+                        // Display code implementation on first page only
+                        CodeImplementationColumn(modifier = Modifier.padding(horizontal = dimensionResource(id = com.orange.ods.R.dimen.screen_horizontal_margin))) {
+                            FunctionCallCode(
+                                name = if (scrollableTabs) OdsComposable.OdsScrollableTabRow.name else OdsComposable.OdsTabRow.name,
+                                parameters = {
+                                    int("selectedTabIndex", pageIndex)
+                                    list("tabs") {
+                                        tabs.forEach { tab ->
+                                            classInstance<OdsTabRow.Tab> {
+                                                if (tabIconEnabled.value) {
+                                                    classInstance<OdsTabRow.Tab.Icon>("icon") {
+                                                        painter()
+                                                    }
+                                                }
+                                                if (tabTextEnabled.value) {
+                                                    text(tab.name)
+                                                }
+                                                onClick()
+                                            }
+                                        }
+                                    }
+                                    enum("tabIconPosition", tabIconPosition.value)
+                                }
+                            )
+                        }
+                    } else {
+                        OdsText(text = stringResource(tabs[pageIndex].textResId), style = OdsTextStyle.BodyL)
+                    }
+                }
             }
         }
-    }
-}
-
-@Composable
-private fun TabsPagerContentScreen(text: String) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .wrapContentSize(Alignment.Center)
-    ) {
-        OdsText(text = text, style = OdsTextStyle.BodyL)
     }
 }